Handle indentations by preparsing token stream
Originally the lexer emits extra INDENT, DEDENT and NEWLINE tokens
for the parser to recognise the whitespace-sensitive grammar. This
approach is based on Python and it is relatively simple to implement.

However, this approach makes parsing certain constructs (eg multiline
expressions) unnecessarily difficult for a LR(1) parser, where the
grammar would be more natural with two lookahead tokens.

Starting from this commit we switch to a preparser-based approach
(based on F#). The token stream is augmented with artificial tokens
such as BEGIN, END, IN, etc to delimit syntactic constructs.